long long int for Nbytes

+++ src/functions.cpp
@@ -358,7 +358,7 @@ arma::mat genotypeMatrix(const std::string fileName, int N, int P,
   int i = 0;
   int ii = 0;
   const bool colskip = (col_skip_pos.n_elem > 0);
-  int Nbytes = ceil(N / 4.0);
+  unsigned long long int Nbytes = ceil(N / 4.0);
   const bool selectrow = (keepbytes.n_elem > 0);
   int n, p, nskip;
   if (selectrow)
